Generalize FunctionGraph for both Transformation and Combination
